@@ -10,17 +10,17 @@ import Tags from './Pages/Tags/Tags'
 import Users from './Pages/Users/Users'
 import UserProfile from './Pages/UserProfile/UserProfile'
 
-const AllRoutes = () => {
+const AllRoutes = ({ slideIn, handleSlideIn }) => {
     return (
         <Routes>
-            <Route path='/' element={<Home />}/>
+            <Route path='/' element={<Home slideIn={slideIn} handleSlideIn={handleSlideIn}/>}/>
             <Route path='/Auth' element={<Auth />}/>
             <Route path='/AskQuestion' element={<AskQuestion />}/>
-            <Route path='/Questions' element={<Questions />}/>
-            <Route path='/Questions/:id' element={<DisplayQuestion />}/>
-            <Route path='/Tags' element={<Tags />} />
-            <Route path='/Users' element={<Users />} />
-            <Route path='/Users/:id' element={<UserProfile />} />
+            <Route path='/Questions' element={<Questions slideIn={slideIn} handleSlideIn={handleSlideIn}/>}/>
+            <Route path='/Questions/:id' element={<DisplayQuestion slideIn={slideIn} handleSlideIn={handleSlideIn}/>}/>
+            <Route path='/Tags' element={<Tags slideIn={slideIn} handleSlideIn={handleSlideIn}/>} />
+            <Route path='/Users' element={<Users slideIn={slideIn} handleSlideIn={handleSlideIn}/>} />
+            <Route path='/Users/:id' element={<UserProfile slideIn={slideIn} handleSlideIn={handleSlideIn}/>} />
         </Routes>
     )
 }
