@@ -1,5 +1,5 @@
 import { BrowserRouter as Router } from 'react-router-dom' 
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import './App.css';
 import Navbar from './components/Navbar/Navbar'
@@ -16,11 +16,22 @@ function App() {
     dispatch(fetchAllUsers())
   }, [dispatch])
 
+  
+  const [ slideIn, setSlideIn ] = useState(true)
+  
+  useEffect(()=>{
+    if(window.innerWidth <= 760){
+      setSlideIn(false)
+    }
+  }, [])
+
+  const handleSlideIn = () => setSlideIn(state => !state)
+
   return (
     <div className="App">
       <Router >
-        <Navbar />
-        <AllRoutes />
+        <Navbar handleSlideIn = {handleSlideIn}/>
+        <AllRoutes slideIn={slideIn} handleSlideIn={handleSlideIn}/>
       </Router >
     </div>
   );
