@@ -16,8 +16,13 @@ const AskQuestion = () => {
 
     const handleSubmit = (e) => {
         e.preventDefault()
-        // console.log({ questionTitle, questionBody, questionTags})
-        dispatch(askQuestion({ questionTitle, questionBody, questionTags, userPosted: User.result.name }, navigate))
+        if(User){
+            if(questionTitle && questionBody && questionTags){
+                dispatch(askQuestion({ questionTitle, questionBody, questionTags, userPosted: User.result.name }, navigate))
+            }
+            else alert("Please enter all the fields")
+        }
+        else alert("Login to ask question");
     }
 
     const handleEnter = (e) => {
