@@ -5,10 +5,10 @@ import LeftSidebar from '../../components/LeftSidebar/LeftSidebar'
 import RightSidebar from '../../components/RightSidebar/RightSidebar'
 import HomeMainbar from '../../components/HomeMainbar/HomeMainbar'
 
-const Home = () => {
+const Home = ({slideIn}) => {
     return (
         <div className='home-container-1'>
-            <LeftSidebar />
+            <LeftSidebar slideIn={slideIn}/>
             <div className='home-container-2'>
                 <HomeMainbar />
                 <RightSidebar />
