@@ -3,10 +3,10 @@ import LeftSidebar from '../../components/LeftSidebar/LeftSidebar'
 import RightSidebar from '../../components/RightSidebar/RightSidebar'
 import QuestionsDetails from './QuestionsDetails'
 
-const DisplayQuestion = () => {
+const DisplayQuestion = ({slideIn, handleSlideIn}) => {
     return (
         <div className='home-container-1'>
-            <LeftSidebar />
+            <LeftSidebar slideIn={slideIn} handleSlideIn={handleSlideIn}/>
             <div className='home-container-2'>
                 <QuestionsDetails />
                 <RightSidebar />
