@@ -5,10 +5,10 @@ import LeftSidebar from '../../components/LeftSidebar/LeftSidebar'
 import RightSidebar from '../../components/RightSidebar/RightSidebar'
 import HomeMainbar from '../../components/HomeMainbar/HomeMainbar'
 
-const Questions = () => {
+const Questions = ({slideIn, handleSlideIn}) => {
     return (
         <div className='home-container-1'>
-            <LeftSidebar />
+            <LeftSidebar slideIn={slideIn} handleSlideIn={handleSlideIn}/>
             <div className='home-container-2'>
                 <HomeMainbar />
                 <RightSidebar />
