@@ -4,11 +4,11 @@ import './Users.css'
 import LeftSidebar from '../../components/LeftSidebar/LeftSidebar'
 import UsersList from './UsersList'
 
-const Users = () => {
+const Users = ({slideIn, handleSlideIn}) => {
 
     return (
         <div className='home-container-1'>
-            <LeftSidebar />
+            <LeftSidebar slideIn={slideIn} handleSlideIn={handleSlideIn}/>
             <div className="home-container-2" style={{marginTop: "30px"}}>
                 <h1 style={{fontWeight: "400"}}>Users</h1>
                 <UsersList /> 
