@@ -16,7 +16,7 @@ export const AskQuestion = async (req, res) => {
 
 export const getAllQuestions = async (req, res) => {
     try {
-        const questionList = await Questions.find();
+        const questionList = await Questions.find().sort({"askedOn": -1});
         res.status(200).json(questionList);
     } catch (error) {
         res.status(404).json({ message: error.message });
